realm_provider: refactor
