+++ integration_test/git_flow_test.dart
@@ -9,7 +9,7 @@ import 'package:pwd/common/domain/usecases/hash_usecase.dart';
 import 'package:pwd/home/presentation/home_tabbar/home_screen.dart';
 import 'package:pwd/main.dart' as app;
 import 'package:pwd/notes/data/datasource/realm_datasource/realm_local_repository_impl.dart';
-import 'package:pwd/notes/data/datasource/realm_datasource/realm_provider_impl.dart';
+import 'package:pwd/notes/data/datasource/realm_datasource/realm_provider/realm_provider_impl.dart';
 
 import 'package:pwd/notes/presentation/edit_note/edit_note_screen.dart';
 import 'pages/configurations_screen/configurations_screen_robot.dart';
