+++ lib/notes/presentation/di/google_and_realm_di.dart
@@ -1,7 +1,7 @@
 import 'package:di_storage/di_storage.dart';
 import 'package:pwd/notes/data/datasource/checksum_checker_impl.dart';
 import 'package:pwd/notes/data/datasource/realm_datasource/realm_local_repository_impl.dart';
-import 'package:pwd/notes/data/datasource/realm_datasource/realm_provider_impl.dart';
+import 'package:pwd/notes/data/datasource/realm_datasource/realm_provider/realm_provider_impl.dart';
 import 'package:pwd/notes/domain/checksum_checker.dart';
 import 'package:pwd/notes/domain/google_repository.dart';
